@@ -15,14 +15,22 @@
 		}
 	],
 	"require": {
-		"php": "^7.2|^8.0",
-		"illuminate/support": "~5.5.0|~5.6.0|~5.7.0|~5.8.0|^6.0|^7.0|^8.0"
+		"php": "^8.0.2|^8.1|^8.2",
+		"illuminate/support": "^8.0|^9.0|^10.0"
 	},
 	"require-dev": {
+		"illuminate/filesystem": "^8.0|^9.0|^10.0",
+		"laravel/framework": "^8.0|^9.0|^10.0",
+		"laravel/pint": "^1.0",
 		"mockery/mockery": "^1.0.0",
-		"illuminate/filesystem": "~5.5.0|~5.6.0|~5.7.0|~5.8.0|^6.0|^7.0|^8.0|^9.0",
-		"laravel/framework": "~5.5.0|~5.6.0|~5.7.0|~5.8.0|^6.0|^7.0|^8.0|^9.0",
-		"phpunit/phpunit": "^8.0"
+		"nunomaduro/collision": "^6.0",
+		"nunomaduro/larastan": "^2.0.1",
+		"orchestra/testbench": "^7.0|^8.0",
+		"phpstan/extension-installer": "^1.1",
+		"phpstan/phpstan-deprecation-rules": "^1.0",
+		"phpstan/phpstan-phpunit": "^1.0",
+		"phpunit/phpunit": "^9.5",
+		"rector/rector": "^0.15.17"
 	},
 	"autoload": {
 		"psr-0": {
@@ -37,6 +45,12 @@
 			"Orangehill\\Iseed\\Tests\\": "tests"
 		}
 	},
+	"scripts": {
+		"analyse": "vendor/bin/phpstan analyse",
+		"test": "vendor/bin/phpunit",
+		"test-coverage": "vendor/bin/phpunit --coverage",
+		"format": "vendor/bin/pint"
+	},
 	"extra": {
 		"laravel": {
 			"providers": [
@@ -45,7 +59,10 @@
 		}
 	},
 	"config": {
-		"sort-packages": true
+		"sort-packages": true,
+		"allow-plugins": {
+			"phpstan/extension-installer": true
+		}
 	},
 	"minimum-stability": "dev",
 	"prefer-stable": true
