@@ -8,70 +8,56 @@ class IseedServiceProvider extends ServiceProvider
 {
     /**
      * Indicates if loading of the provider is deferred.
-     *
-     * @var bool
      */
-    protected $defer = false;
+    protected bool $defer = false;
 
     /**
      * Bootstrap the application events.
-     *
-     * @return void
      */
-    public function boot()
+    public function boot(): void
     {
         require base_path().'/vendor/autoload.php';
     }
 
     /**
      * Register the service provider.
-     *
-     * @return void
      */
-    public function register()
+    public function register(): void
     {
         $this->registerResources();
 
-        $this->app->singleton('iseed', function($app) {
-            return new Iseed;
-        });
+        $this->app->singleton('iseed', fn ($app) => new Iseed);
 
-        $this->app->booting(function() {
+        $this->app->booting(function () {
             $loader = \Illuminate\Foundation\AliasLoader::getInstance();
-            $loader->alias('Iseed', 'Orangehill\Iseed\Facades\Iseed');
+            $loader->alias('Iseed', \Orangehill\Iseed\Facades\Iseed::class);
         });
 
-        $this->app->singleton('command.iseed', function($app) {
-            return new IseedCommand;
-        });
+        $this->app->singleton('command.iseed', fn ($app) => new IseedCommand);
 
         $this->commands('command.iseed');
     }
 
     /**
      * Get the services provided by the provider.
-     *
-     * @return array
      */
-    public function provides()
+    public function provides(): array
     {
-        return array('iseed');
+        return ['iseed'];
     }
 
     /**
      * Register the package resources.
-     *
-     * @return void
      */
-    protected function registerResources()
+    protected function registerResources(): void
     {
-        $userConfigFile    = app()->configPath().'/iseed.php';
+        $userConfigFile = app()->configPath().'/iseed.php';
         $packageConfigFile = __DIR__.'/../../config/config.php';
-        $config            = $this->app['files']->getRequire($packageConfigFile);
+        $config = $this->app['files']->getRequire($packageConfigFile);
 
         if (file_exists($userConfigFile)) {
             $userConfig = $this->app['files']->getRequire($userConfigFile);
-            $config     = array_replace_recursive($config, $userConfig);
+            $config = array_replace_recursive($config, $userConfig);
         }
 
         $this->app['config']->set('iseed::config', $config);
