@@ -3,6 +3,9 @@
 namespace Orangehill\Iseed;
 
 use Illuminate\Console\Command;
+use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Schema;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 
@@ -23,45 +26,33 @@ class IseedCommand extends Command
     protected $description = 'Generate seed file from table';
 
     /**
-     * Create a new command instance.
-     *
-     * @return \Orangehill\Iseed\IseedCommand
-     */
-    public function __construct()
-    {
-        parent::__construct();
-    }
-
-    /**
      * Execute the console command.
-     *
-     * @return void
      */
-    public function handle()
+    public function handle(): int
     {
-        return $this->fire();
+        $this->fire();
+
+        return 1;
     }
 
     /**
      * Execute the console command (for <= 5.4).
-     *
-     * @return void
      */
-    public function fire()
+    public function fire(): void
     {
         // if clean option is checked empty iSeed template in DatabaseSeeder.php
         if ($this->option('clean')) {
             app('iseed')->cleanSection();
         }
 
-        $tables = explode(",", $this->argument('tables'));
+        $tables = explode(',', $this->argument('tables'));
         $max = intval($this->option('max'));
         $chunkSize = intval($this->option('chunksize'));
-        $exclude = explode(",", $this->option('exclude'));
-        $prerunEvents = explode(",", $this->option('prerun'));
-        $postrunEvents = explode(",", $this->option('postrun'));
+        $exclude = explode(',', $this->option('exclude'));
+        $prerunEvents = explode(',', $this->option('prerun'));
+        $postrunEvents = explode(',', $this->option('postrun'));
         $dumpAuto = intval($this->option('dumpauto'));
-        $indexed = !$this->option('noindex');
+        $indexed = ! $this->option('noindex');
         $orderBy = $this->option('orderby');
         $direction = $this->option('direction');
         $prefix = $this->option('classnameprefix');
@@ -89,10 +80,10 @@ class IseedCommand extends Command
             $tableIncrement++;
 
             // generate file and class name based on name of the table
-            list($fileName, $className) = $this->generateFileName($table, $prefix, $suffix);
+            [$fileName, $className] = $this->generateFileName($table, $prefix, $suffix);
 
             // if file does not exist or force option is turned on generate seeder
-            if (!\File::exists($fileName) || $this->option('force')) {
+            if (! File::exists($fileName) || $this->option('force')) {
                 $this->printResult(
                     app('iseed')->generateSeed(
                         $table,
@@ -111,10 +102,11 @@ class IseedCommand extends Command
                     ),
                     $table
                 );
+
                 continue;
             }
 
-            if ($this->confirm('File ' . $className . ' already exist. Do you wish to override it? [yes|no]')) {
+            if ($this->confirm('File '.$className.' already exist. Do you wish to override it? [yes|no]')) {
                 // if user said yes overwrite old seeder
                 $this->printResult(
                     app('iseed')->generateSeed(
@@ -134,58 +126,52 @@ class IseedCommand extends Command
                 );
             }
         }
-
-        return;
     }
 
     /**
      * Get the console command arguments.
-     *
-     * @return array
      */
-    protected function getArguments()
+    protected function getArguments(): array
     {
-        return array(
-            array('tables', InputArgument::REQUIRED, 'comma separated string of table names'),
-        );
+        return [
+            ['tables', InputArgument::REQUIRED, 'comma separated string of table names'],
+        ];
     }
 
     /**
      * Get the console command options.
-     *
-     * @return array
      */
-    protected function getOptions()
+    protected function getOptions(): array
     {
-        return array(
-            array('clean', null, InputOption::VALUE_NONE, 'clean iseed section', null),
-            array('force', null, InputOption::VALUE_NONE, 'force overwrite of all existing seed classes', null),
-            array('database', null, InputOption::VALUE_OPTIONAL, 'database connection', \Config::get('database.default')),
-            array('max', null, InputOption::VALUE_OPTIONAL, 'max number of rows', null),
-            array('chunksize', null, InputOption::VALUE_OPTIONAL, 'size of data chunks for each insert query', null),
-            array('exclude', null, InputOption::VALUE_OPTIONAL, 'exclude columns', null),
-            array('prerun', null, InputOption::VALUE_OPTIONAL, 'prerun event name', null),
-            array('postrun', null, InputOption::VALUE_OPTIONAL, 'postrun event name', null),
-            array('dumpauto', null, InputOption::VALUE_OPTIONAL, 'run composer dump-autoload', true),
-            array('noindex', null, InputOption::VALUE_NONE, 'no indexing in the seed', null),
-            array('orderby', null, InputOption::VALUE_OPTIONAL, 'orderby desc by column', null),
-            array('direction', null, InputOption::VALUE_OPTIONAL, 'orderby direction', null),
-            array('classnameprefix', null, InputOption::VALUE_OPTIONAL, 'prefix for class and file name', null),
-            array('classnamesuffix', null, InputOption::VALUE_OPTIONAL, 'suffix for class and file name', null),
-        );
+        return [
+            ['clean', null, InputOption::VALUE_NONE, 'clean iseed section', null],
+            ['force', null, InputOption::VALUE_NONE, 'force overwrite of all existing seed classes', null],
+            ['database', null, InputOption::VALUE_OPTIONAL, 'database connection', Config::get('database.default')],
+            ['max', null, InputOption::VALUE_OPTIONAL, 'max number of rows', null],
+            ['chunksize', null, InputOption::VALUE_OPTIONAL, 'size of data chunks for each insert query', null],
+            ['exclude', null, InputOption::VALUE_OPTIONAL, 'exclude columns', null],
+            ['prerun', null, InputOption::VALUE_OPTIONAL, 'prerun event name', null],
+            ['postrun', null, InputOption::VALUE_OPTIONAL, 'postrun event name', null],
+            ['dumpauto', null, InputOption::VALUE_OPTIONAL, 'run composer dump-autoload', true],
+            ['noindex', null, InputOption::VALUE_NONE, 'no indexing in the seed', null],
+            ['orderby', null, InputOption::VALUE_OPTIONAL, 'orderby desc by column', null],
+            ['direction', null, InputOption::VALUE_OPTIONAL, 'orderby direction', null],
+            ['classnameprefix', null, InputOption::VALUE_OPTIONAL, 'prefix for class and file name', null],
+            ['classnamesuffix', null, InputOption::VALUE_OPTIONAL, 'suffix for class and file name', null],
+        ];
     }
 
     /**
      * Provide user feedback, based on success or not.
      *
-     * @param  boolean $successful
-     * @param  string $table
+     * @param  bool  $successful
      * @return void
      */
-    protected function printResult($successful, $table)
+    protected function printResult(bool|int $successful, string $table)
     {
         if ($successful) {
             $this->info("Created a seed file from table {$table}");
+
             return;
         }
 
@@ -195,18 +181,19 @@ class IseedCommand extends Command
     /**
      * Generate file name, to be used in test wether seed file already exist
      *
-     * @param  string $table
+     *
      * @return string
      */
-    protected function generateFileName($table, $prefix=null, $suffix=null)
+    protected function generateFileName(string $table, ?string $prefix = null, ?string $suffix = null): array|string
     {
-        if (!\Schema::connection($this->option('database') ? $this->option('database') : config('database.default'))->hasTable($table)) {
+        if (! Schema::connection($this->option('database') ?: config('database.default'))->hasTable($table)) {
             throw new TableNotFoundException("Table $table was not found.");
         }
 
         // Generate class name and file name
         $className = app('iseed')->generateClassName($table, $prefix, $suffix);
-        $seedPath = base_path() . config('iseed::config.path');
-        return [$seedPath . '/' . $className . '.php', $className . '.php'];
+        $seedPath = base_path().config('iseed::config.path');
+
+        return [$seedPath.'/'.$className.'.php', $className.'.php'];
     }
 }
