@@ -1,5 +1,6 @@
 <?php
-return array(
+
+return [
     'path' => '/database/seeders',
-    'chunk_size' => 500 // Maximum number of rows per insert statement
-);
+    'chunk_size' => 500, // Maximum number of rows per insert statement
+];
